add branch behaviortree

+++ scripts/make_types.sh
@@ -17,7 +17,12 @@ rm */*.pyc
 
 # Make
 lcm-gen -jxp *.lcm
-cp /usr/local/share/java/lcm.jar .
+if [ -f "/usr/local/share/java/lcm.jar" ]; then
+  cp /usr/local/share/java/lcm.jar .
+else
+  cp /usr/share/java/lcm.jar .
+fi
+cp /usr/share/java/lcm.jar .
 javac -cp lcm.jar */*.java
 jar cf my_types.jar */*.class
 mkdir -p java
